requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
faster-whisper
python-docx
pydub
opencc-python-reimplemented
anyio[trio]
websockets
pytest==8.2.0
httpx==0.27.0
